requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cloudscraper==1.2.58
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
logzero==1.7.0
multidict==5.2.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
protobuf==3.19.1
py-cord==2.0.0a4580+g1d65214e
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyotp==2.6.0
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
